datablobstorage: put mock functions in testutil.go
Just a simple movement of codes.

Release note: None
